Added jsdoc to error
