+++ src/capitalize-first-letter.js
