--- src/error/error-not-found.mjs
@@ -1,11 +1,0 @@
-class ErrorNotFound extends Error {
-    constructor(message) {
-        super(message);
-
-        this.name = this.constructor.name;
-
-        Error.captureStackTrace(this, this.constructor);
-    }
-}
-
-export {ErrorNotFound};
