+++ src/index.js
@@ -1,2 +1,2 @@
 export {KubeConfig} from '@kubernetes/client-node';
-export * from './k8s-client.mjs';
+export * from './k8s-client.js';
