+++ src/normalize-kind.js
